requirements.txt
# Streamlit Cloud compatible requirements
streamlit>=1.47.0
openai>=1.3.0
faiss-cpu>=1.10.0

# Document processing
pdfplumber>=0.11.0
pypdf>=4.2.0
python-docx>=1.1.0

# Data processing
pandas>=2.1.0
openpyxl>=3.1.0
xlrd>=2.0.0
numpy>=1.26.0

# Reporting
reportlab>=3.6.0

# Environment and HTTP
python-dotenv>=1.0.0
httpx>=0.24.0,<0.28.0

# Additional dependencies that might be missing
typing-extensions>=4.0.0
cryptography>=41.0.0
